move testdata into __tests__
