+++ src/components/__tests__/App.test.js
@@ -7,7 +7,7 @@ import injectTapEventPlugin from 'react-tap-event-plugin';
 import App from '../App';
 import TopBar from '../TopBar';
 import SideBar from '../SideBar';
-import { testBooks, jsonHeaders } from '../../common/testData';
+import { testBooks, jsonHeaders } from '../../common/__tests__/testData';
 
 describe('App', () => {
 	let wrapper;
